Move detector to another folder
