+++ slither/detectors/statements/unused_custom_errors.py
